rename: actions 디렉토리명을 server-actions 디렉토리명으로 변경 (#87)
